+++ tests/tests_pytorch/deprecated_api/test_remove_1-7.py
@@ -15,14 +15,11 @@
 import os
 from re import escape
 from unittest import mock
-from unittest.mock import Mock
 
 import pytest
 import torch
 
 from pytorch_lightning import Trainer
-from pytorch_lightning.demos.boring_classes import BoringModel
-from pytorch_lightning.overrides.distributed import IndexBatchSamplerWrapper
 from pytorch_lightning.plugins.environments import (
     KubeflowEnvironment,
     LightningEnvironment,
@@ -32,27 +29,6 @@ from pytorch_lightning.plugins.environments import (
 )
 from pytorch_lightning.strategies import SingleDeviceStrategy
 from tests_pytorch.plugins.environments.test_lsf_environment import _make_rankfile
-
-
-def test_v1_7_0_deprecate_on_post_move_to_device(tmpdir):
-    class TestModel(BoringModel):
-        def on_post_move_to_device(self):
-            print("on_post_move_to_device")
-
-    model = TestModel()
-
-    trainer = Trainer(default_root_dir=tmpdir, limit_train_batches=5, max_epochs=1)
-
-    with pytest.deprecated_call(
-        match=r"Method `on_post_move_to_device` has been deprecated in v1.5 and will be removed in v1.7"
-    ):
-        trainer.fit(model)
-
-
-def test_v1_7_0_deprecated_slurm_job_id():
-    trainer = Trainer()
-    with pytest.deprecated_call(match="Method `slurm_job_id` is deprecated in v1.6.0 and will be removed in v1.7.0."):
-        trainer.slurm_job_id
 
 
 def test_v1_7_0_deprecated_max_steps_none(tmpdir):
@@ -139,15 +115,6 @@ def test_v1_7_0_cluster_environment_detection(cls, method_name, tmp_path):
                 match=f"MyClusterEnvironment.{method_name}` has been deprecated in v1.6 and will be removed in v1.7"
             ):
                 MyClusterEnvironment()
-
-
-def test_v1_7_0_index_batch_sampler_wrapper_batch_indices():
-    sampler = IndexBatchSamplerWrapper(Mock())
-    with pytest.deprecated_call(match="was deprecated in v1.5 and will be removed in v1.7"):
-        _ = sampler.batch_indices
-
-    with pytest.deprecated_call(match="was deprecated in v1.5 and will be removed in v1.7"):
-        sampler.batch_indices = []
 
 
 def test_v1_7_0_post_dispatch_hook():